+++ stocal/tests/test_examples.py
@@ -6,7 +6,7 @@ import unittest
 import sys
 import os
 
-from stocal.tests.test_transitions import TestReactionRule, TestMassAction
+from stocal.tests.test_transitions import TestReactionRule as TestTransitionRule, TestMassAction
 
 from stocal.examples.pre2017 import DegradationRule
 from stocal.examples.pre2017 import LigationRule
@@ -40,7 +40,7 @@ class TestPre2017(unittest.TestCase):
             pass
 
 
-class TestPre2017Rule(TestReactionRule):
+class TestPre2017Rule(TestTransitionRule):
     """Base class for rules used in pre2017"""
     def setUp(self):
         self.rule = self.Rule()
@@ -119,7 +119,7 @@ class TestPre2017AutoCatalysis(TestMassAction):
                 reactants, products, c)
 
 
-class TestTypedRules(TestReactionRule):
+class TestTypedRules(TestTransitionRule):
     from stocal.examples.typed_rules import AA_BB as Rule
 
     def test_infer_transitions_signature(self):